Fix egold checkout param

+++ features/digital/common_digital/src/main/java/com/tokopedia/common_digital/cart/data/entity/requestbody/checkout/FintechProductCheckout.kt
@@ -18,7 +18,7 @@ class FintechProductCheckout (
     var tierId: Int = 0,
     @SerializedName("user_id")
     @Expose
-    var userId: Int? = null,
+    var userId: Long? = null,
     @SerializedName("fintech_amount")
     @Expose
     var fintechAmount: Long = 0,
